+++ frontend/src/metabase/components/BrowserCrumbs.jsx
@@ -6,27 +6,27 @@ import Subhead from "metabase/components/Subhead";
 
 // TODO: merge with Breadcrumbs
 
-const BrowseHeader = ({ children }) => (
-  <Box my={3}>
-    <Subhead>{children}</Subhead>
-  </Box>
-);
+const BrowseHeader = ({ children }) => <Subhead>{children}</Subhead>;
 
 const BrowserCrumbs = ({ crumbs }) => (
-  <Flex align="center">
+  <Box>
     {crumbs.filter(c => c).map((crumb, index, crumbs) => [
       crumb.to ? (
-        <Link key={"title" + index} to={crumb.to}>
-          <BrowseHeader>{crumb.title}</BrowseHeader>
-        </Link>
+        <Flex align="center">
+          <Link key={"title" + index} to={crumb.to}>
+            {crumb.title}
+          </Link>
+          {index < crumbs.length - 1 ? (
+            <Icon key={"divider" + index} name="chevronright" mx={1} />
+          ) : null}
+        </Flex>
       ) : (
-        <BrowseHeader>{crumb.title}</BrowseHeader>
+        <Box>
+          <BrowseHeader>{crumb.title}</BrowseHeader>
+        </Box>
       ),
-      index < crumbs.length - 1 ? (
-        <Icon key={"divider" + index} name="chevronright" mx={2} />
-      ) : null,
     ])}
-  </Flex>
+  </Box>
 );
 
 export default BrowserCrumbs;
