+++ frontend/src/metabase/components/CollectionLanding.jsx
@@ -20,7 +20,6 @@ import Icon from "metabase/components/Icon";
 import Link from "metabase/components/Link";
 import CollectionEmptyState from "metabase/components/CollectionEmptyState";
 import EntityMenu from "metabase/components/EntityMenu";
-import Ellipsified from "metabase/components/Ellipsified";
 import VirtualizedList from "metabase/components/VirtualizedList";
 import BrowserCrumbs from "metabase/components/BrowserCrumbs";
 
@@ -29,88 +28,14 @@ import { entityObjectLoader } from "metabase/entities/containers/EntityObjectLoa
 
 import { ROOT_COLLECTION } from "metabase/entities/collections";
 
+import CollectionList from "metabase/components/CollectionList";
+
 // drag-and-drop components
 import ItemDragSource from "metabase/containers/dnd/ItemDragSource";
 import CollectionDropTarget from "metabase/containers/dnd/CollectionDropTarget";
 import PinPositionDropTarget from "metabase/containers/dnd/PinPositionDropTarget";
 import PinDropTarget from "metabase/containers/dnd/PinDropTarget";
 import ItemsDragLayer from "metabase/containers/dnd/ItemsDragLayer";
-
-const CollectionItem = ({ collection, color, iconName = "all" }) => (
-  <Link
-    to={`collection/${collection.id}`}
-    hover={{ color: normal.blue }}
-    color={color || normal.grey2}
-  >
-    <Card p={1}>
-      <Flex align="center" py={1} key={`collection-${collection.id}`}>
-        <Icon name={iconName} mx={1} color="#93B3C9" />
-        <h4>
-          <Ellipsified>{collection.name}</Ellipsified>
-        </h4>
-      </Flex>
-    </Card>
-  </Link>
-);
-
-@connect(({ currentUser }) => ({ currentUser }), null)
-class CollectionList extends React.Component {
-  render() {
-    const { collections, currentUser, isRoot } = this.props;
-    return (
-      <Box mb={2}>
-        <Grid>
-          {isRoot && (
-            <GridItem w={1 / 4} className="relative">
-              <CollectionDropTarget
-                collection={{ id: currentUser.personal_collection_id }}
-              >
-                <CollectionItem
-                  collection={{
-                    name: t`My personal collection`,
-                    id: currentUser.personal_collection_id,
-                  }}
-                  iconName="star"
-                />
-              </CollectionDropTarget>
-            </GridItem>
-          )}
-          {isRoot &&
-            currentUser.is_superuser && (
-              <GridItem w={1 / 4}>
-                <CollectionItem
-                  collection={{
-                    name: t`Everyone else's personal collections`,
-                    // Bit of a hack. The route /collection/users lists
-                    // user collections but is not itself a colllection,
-                    // but using the fake id users here works
-                    id: "users",
-                  }}
-                  iconName="person"
-                />
-              </GridItem>
-            )}
-          {collections
-            .filter(c => c.id !== currentUser.personal_collection_id)
-            .map(collection => (
-              <GridItem
-                w={1 / 4}
-                key={collection.id}
-                mb={1}
-                className="relative"
-              >
-                <CollectionDropTarget collection={collection}>
-                  <ItemDragSource item={collection}>
-                    <CollectionItem collection={collection} />
-                  </ItemDragSource>
-                </CollectionDropTarget>
-              </GridItem>
-            ))}
-        </Grid>
-      </Box>
-    );
-  }
-}
 
 const ROW_HEIGHT = 72;
 
@@ -196,7 +121,7 @@ class DefaultLanding extends React.Component {
                     >
                       <Grid>
                         {pinned.map((item, index) => (
-                          <GridItem w={1 / 3} className="relative">
+                          <GridItem w={1 / 4} className="relative">
                             <ItemDragSource item={item}>
                               <PinnedItem
                                 key={`${item.type}:${item.id}`}
@@ -216,7 +141,7 @@ class DefaultLanding extends React.Component {
                           </GridItem>
                         ))}
                         {pinned.length % 2 === 1 ? (
-                          <GridItem w={1 / 3} className="relative">
+                          <GridItem w={1 / 4} className="relative">
                             <PinPositionDropTarget
                               pinIndex={
                                 pinned[pinned.length - 1].collection_position +
@@ -489,7 +414,7 @@ class CollectionLanding extends React.Component {
     return (
       <Box mx={4}>
         <Box>
-          <Flex align="center">
+          <Flex align="center" my={2}>
             <BrowserCrumbs
               crumbs={[
                 ...ancestors.map(({ id, name }) => ({
